api/requirements.txt
fastapi
uvicorn
pandas
mlflow
psycopg2-binary
scikit-learn
